Rework cli file structure
