+++ index.ts
@@ -4,7 +4,7 @@ export {
   UniqueConstraintError,
   OnlyOneError,
 } from "./errors.js";
-export { cli } from "./cli.js";
+export { cli } from "./cli/index.js";
 
 export type { PgistConfig, TxFn, Queryable } from "./db.js";
 export type { QueryResult } from "./query.js";
